+++ samples/SampleWebApi/Startup.cs
@@ -1,10 +1,15 @@
-﻿using FluentValidation.AspNetCore;
+﻿using System.Linq;
+using FluentValidation;
+using FluentValidation.AspNetCore;
+using MicroElements.Swashbuckle.FluentValidation;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using SampleWebApi.Validators;
+using SampleWebApi.DbModels;
 using Swashbuckle.AspNetCore.Swagger;
 
 namespace SampleWebApi
@@ -24,7 +29,19 @@ namespace SampleWebApi
             services
                 .AddMvc()
                 // Adds fluent validators to Asp.net
-                .AddFluentValidation(fv => fv.RegisterValidatorsFromAssemblyContaining<CustomerValidator>());
+                .AddFluentValidation(c =>
+                {
+                    c.RegisterValidatorsFromAssemblyContaining<Startup>();
+                    // Optionally set validator factory if you have problems with scope resolve inside validators.
+                    // c.ValidatorFactoryType = typeof(CustomValidatorFactory);
+                });
+
+            // Register all validators as IValidator?
+            var serviceDescriptors = services.Where(descriptor => descriptor.ServiceType.GetInterfaces().Contains(typeof(IValidator))).ToList();
+            serviceDescriptors.ForEach(descriptor => services.Add(ServiceDescriptor.Transient(typeof(IValidator), descriptor.ImplementationType)));
+
+            // One more way to set custom factory.
+            //services = services.Replace(ServiceDescriptor.Scoped<IValidatorFactory, ScopedServiceProviderValidatorFactory>());
 
             services.AddSwaggerGen(c =>
             {
@@ -35,6 +52,13 @@ namespace SampleWebApi
 
             // Adds logging
             services.AddLogging(builder => builder.AddConsole());
+
+            // Register database
+            services.AddDbContext<BloggingDbContext>(
+                options => options.UseInMemoryDatabase("validationDB"),
+                ServiceLifetime.Scoped);
+
+            //services.AddTransient<Func<BloggingDbContext>>(provider => provider.GetService<BloggingDbContext>);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -43,13 +67,38 @@ namespace SampleWebApi
             app
                 .UseMvc()
                 // Adds swagger
-                .UseSwagger();
+                //.UseSwagger()
+
+                // Use scoped swagger if you have problems with scoped services in validators
+                .UseScopedSwagger();
+            ;
 
             // Adds swagger UI
             app.UseSwaggerUI(c =>
             {
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1");
             });
+
+            // Seed database (NOT PRODUCTION CODE)
+            TestDatabaseSeed(app);
+        }
+
+        private static void TestDatabaseSeed(IApplicationBuilder app)
+        {
+            var bloggingContext = app.ApplicationServices.CreateScope().ServiceProvider.GetService<BloggingDbContext>();
+            if (!EnumerableExtensions.Any(bloggingContext.Metadata))
+            {
+                // Example of defining rules dynamically from database.
+                bloggingContext.Metadata.Add(new ValidationMetadata
+                {
+                    ValidationMetadataId = 1,
+                    TypeName = typeof(Blog).Name,
+                    PropertyName = nameof(Blog.Author),
+                    IsRequired = true,
+                    MaxLength = 7
+                });
+                bloggingContext.SaveChanges();
+            }
         }
     }
 }