+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes in 1.0.0:
+* Added ScopedSwaggerMiddleware to resolve error "Cannot resolve 'IValidator<T>' from root provider because it requires scoped service 'TDependency'"
 * Updated samples and documentation
-* Build scripts replaced with MicroElements.Devops
+* Build scripts migrated to MicroElements.Devops
 * Build: added SourceLink
 
 # Changes in 0.8.2: